% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create Graph of Accidents by State}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{id of state of interest}

\item{year}{Year to plot accidents for.}
}
\value{
Returns a state map for the state.num provided for the year provided, with a point in the geographic location of each accident that occured.
}
\description{
This function returns a plot of all accidents in a state for the year provided.
An error will occur if invalid state.num provided.
If no accidents exist for state.num and year provided, a message will be reutrned stating this.
}
