% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Import File as Tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The filename the function reads.}
}
\value{
The function returns a tbl_df.
}
\description{
This is a simple function that imports a csv file as a tibble dataframe.
The function check if the file exists, and if it does, reads in the file and converts it to a tbl_df.
An error will occur if filename doe not exist.
}
\examples{
fars_read('this_file_does_not_exist.csv')

}
