% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Provide a Summary of Counts for Each Year and Month Comination}
\usage{
fars_summarize_years(years)
}
\value{
A tibble with MONTH as rows, years as columns, and counts of year and month combination as the values.
}
\description{
This funtion returns a tibble for each Year and Month combination that counts the total rows for each month, with each year as its own column.
}
\examples{
fars_summarize_years(c(2013, 2014))

}
