% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make Filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numeric year value.}
}
\value{
Returns a character vector containing a formatted combination of text (eg. "accident_") and numeric year variable supplied.
}
\description{
This function take a year, converts it to an interger, and prints the new file name with the year appended to the end as a character vector (eg. "accident_1996.csv.bz2")
A warning will occur if year is not supplied as a numeric value.
}
\examples{
make_filename(1996)

}
