% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Displays the data in a year file on the map of the corresponding state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{year}{The year for which the accident data is to be displayed.}

\item{state.num}{The state for which the accident data is to be displayed. The value provided must be
an integer representing a state as found in the accident data file.}
}
\value{
A plot of the statement with the accidents of that years.
}
\description{
This function reads every accident file for which the year is provided as input
and displays a state map (state provided as input) with the accidents of that years.
}
\examples{
\dontrun{
fars_map_state(13,2014)
}

}

