% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read filename as a data frame tbl.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file which is to be read. This file must be a valid csv file.}
}
\value{
A data frame tbl of the data contained in filename.
}
\description{
This is a wrapper function to read a file and return it as a data frame tbl.
The function stops if the provided filename does not exist.
If it does exists it reads the file as a csv file without progress bar
 and without returning diagnotstic messages.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}

