% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Makes the data in the year files available as a data frame tbl.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years for which the accident data is to be read.}
}
\value{
A data frame tbl where the rows are month and the columns are the years.
}
\description{
This function reads every accident file for which the year is provided as input
and returns a data frame tbl where the rows are month and the columns are the years.
}
\examples{
\dontrun{
fars_summarize_years(c(2013,2014))
}
}

