% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year which is to be included in the name. This value must be convertible to an integer.}
}
\value{
A filename in the format "accident_<year>.csv.bz2".
}
\description{
This function creates a filename.
}
\examples{
\dontrun{
make_filename(2017)
}
}

